*Created a lifted state in App.js
giftExchangesState as 'lifted' state in the App.js. It is 'lifted' because the app will 'control' it, but pass it and functions that change it to other components.

Added 'hardcoded' prefix to the hardcoded data. Not it is only used to initialize the state.

Added 'State' suffix to  state constant and function returned from useState()